requirements.txt
pillow==12.1.0
pytest>=8.0.0
